Accept floats as inputs

+++ gen_and_run.py
@@ -3,10 +3,10 @@ from sliding_window import create_records_using_window
 
 if __name__ == '__main__':
 
-    window_width = int(input("Enter window width: "))
+    window_width = float(input("Enter window width: "))
 
     create_records_using_window(window_width)
     
     config = setup(window_width)
-    
+
     train(config)
